Address PR feedback

+++ scripts/test/integration-setup.sh
@@ -9,8 +9,8 @@ NEXT_MINOR_VERSION=$(echo "$DBT_VERSION" | awk -F. '{print $1"."$2+1}')
 
 # we install using the following workaround to overcome installation conflicts, such as:
 # apache-airflow 2.3.0 and dbt-core [0.13.0 - 1.5.2] and jinja2>=3.0.0 because these package versions have conflicting dependencies
-rm -f $AIRFLOW_HOME/airflow.cfg
-rm -f $AIRFLOW_HOME/airflow.db
+rm -f "${AIRFLOW_HOME:?}/airflow.cfg"
+rm -f "${AIRFLOW_HOME:?}/airflow.db"
 
 airflow db reset -y
